requirements.txt
boto3
pandas
pyarrow
fastparquet
python-dotenv
psycopg2-binary
google-api-python-client
google-auth
gspread
sqlalchemy
apache-airflow-providers-standard
dbt-snowflake
snowflake-connector-python
apache-airflow-providers-snowflake